Complete work for issue #306
Agent: senior_software_engineer
Task: 403b57da-8a20-4617-9d50-262bb97018e9

[orchestrator-commit]

(Failsafe commit: auto-staged all changes)

+++ src/codetoreum/application/execution_service.py
@@ -107,7 +107,7 @@ class ExecutionService:
         event_store: IEventStore,
         storage: IStorage,
         max_retries: int = 3,
-        retry_delay_seconds: int = 5,
+        retry_delay_seconds: float = 5,
     ):
         """
         Initialize ExecutionService.
